requirements.txt
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
transformers>=4.40.0
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
spacy==3.7.2
boto3==1.34.0
jupyter==1.0.0
matplotlib==3.8.2
seaborn==0.13.0
python-dotenv==1.0.0